fix: 카운트다운 시간 버그 해결
