+++ src/components/common/header/index.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { Badge, Tab, Text } from '@/components/common';
 import SearchBar from './SearchBar';
@@ -6,6 +5,7 @@ import UserNav from './UserNav';
 
 import * as style from './style.css';
 import logo from '@/assets/logo.svg';
+import CountDown from './CountDown';
 
 const TAB_LIST = [
   '실시간 최신 트렌드',
@@ -14,55 +14,7 @@ const TAB_LIST = [
   '크리에이터',
 ];
 
-const calculateCountDown = () => {
-  const today = new Date();
-  const todayTime = today.getTime();
-  const tomorrowTime = new Date(
-    today.getFullYear(),
-    today.getMonth(),
-    today.getDate() + 1,
-    0
-  ).getTime();
-
-  return tomorrowTime - todayTime;
-};
-
-const formatHour = (time: number) => {
-  const hour = Math.ceil((time % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-  return hour.toString().padStart(2, '0');
-};
-
-const formatMin = (time: number) => {
-  const min = Math.ceil((time % (1000 * 60 * 60)) / (1000 * 60));
-  return min.toString().padStart(2, '0');
-};
-
-const formatSec = (time: number) => {
-  const sec = Math.ceil((time % (1000 * 60)) / 1000);
-  return sec.toString().padStart(2, '0');
-};
-
 const Header = () => {
-  const [countDown, setCountDown] = useState(calculateCountDown());
-  const [hour, setHour] = useState('');
-  const [min, setMin] = useState('');
-  const [sec, setSec] = useState('');
-  const timerIdRef = useRef<number>(0);
-
-  useEffect(() => {
-    timerIdRef.current = window.setInterval(() => {
-      setCountDown(calculateCountDown());
-    }, 1000);
-
-    return () => clearInterval(timerIdRef.current);
-  }, []);
-
-  useEffect(() => {
-    setHour(formatHour(countDown));
-    setMin(formatMin(countDown));
-    setSec(formatSec(countDown));
-  }, [countDown]);
-
   return (
     <header className={style.header}>
       <div className={style.top}>
@@ -84,9 +36,7 @@ const Header = () => {
             <Text size="small">오늘의 hypelink</Text>
             <Badge top={-10} right={10} />
           </div>
-          <div className={style.countDown}>
-            {hour}:{min}:{sec}
-          </div>
+          <CountDown />
         </Link>
       </div>
     </header>
